Steam: Small optimizations
git-svn-id: http://svn.miranda-ng.org/main/trunk@11133 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Steam/src/steam_pooling.cpp
@@ -64,7 +64,7 @@ void CSteamProto::ParsePollData(JSONNODE *data)
 				if (status != m_iStatus)
 				{
 					debugLogA("CSteamProto::ParsePollData: Change own status to %i", status);
-					WORD oldStatus = m_iStatus;
+					int oldStatus = m_iStatus;
 					m_iStatus = m_iDesiredStatus = status;
 					ProtoBroadcastAck(NULL, ACKTYPE_STATUS, ACKRESULT_SUCCESS, (HANDLE)oldStatus, m_iStatus);
 				}
